src/requirements.txt
numpy
pytest
Pillow
pyinstaller